arm: mach-snapdragon: Skip USB fixup for qcom,sc7280-dwc3
Currently we unconditionally limits USB to high-speed on all Qualcomm
platforms since most lack super-speed PHY drivers. The sc7280
platform has a working super-speed PHY driver in U-Boot and should
not have this restriction applied.

Add a static check for the "qcom,sc7280-dwc3" compatible string at
the start of fixup_qcom_dwc3() to skip the USB high-speed fixup for
sc7280. Add a log_warning() when the fixup is applied so developers
on other platforms with working super-speed USB know to add their
compatible to the skip list.

Signed-off-by: Balaji Selvanathan <balaji.selvanathan@oss.qualcomm.com>